[MOD/#181] Composable 상태 호이스팅
